# hexabid-api-contract/src/main/resources/openapi/auction-api.yaml
openapi: 3.0.3
info:
  title: Auctions API
  version: 1.0.0
servers:
  - url: http://localhost:18080/hexabid
...
paths:
...
  /api/batches:
    get:
      operationId: browseBatches
      tags:
        - Inventory
      summary: Browse batches for current user
      description: |
        Returns paginated list of production batches owned by the authenticated user.
        Batches are used to track products that come from the same production source.
        
        **Archetypowy kontekst:** Inventory (M06)
        Batch reprezentuje partię produkcyjną - grupę produktów o wspólnym pochodzeniu.
        Np. kontener kukurydzy z gospodarstwa X, zbiór 2026.
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - in: query
          name: productId
          description: Filter by product type ID
          schema:
            type: string
            format: uuid
        - in: query
          name: query
          description: Filter by batch name (partial match)
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - in: query
          name: after
          schema:
            type: string
      responses:
        '200':
          description: Batch list (paginated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchListResponse'
        '401':
          description: Authentication required
    post:
      operationId: createBatch
      tags:
        - Inventory
      summary: Define a new production batch
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
      responses:
        '201':
          description: Batch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '400':
          description: Invalid request
  /api/batches/{batchId}:
    get:
      operationId: getBatch
      tags:
        - Inventory
      summary: Get batch by ID
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - in: path
          name: batchId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Batch details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResponse'
        '404':
          description: Batch not found
  /api/inventory/instances:
    get:
      operationId: browseInventoryInstances
      tags:
        - Inventory
      summary: Browse inventory instances for current user
      description: |
        Returns paginated list of inventory instances owned by the authenticated user.
        Instances represent specific quantities of product that can be sold on auction.
        
        **Archetypowy kontekst:** Inventory (M06) - ProductInstance
        Instancja to konkretna ilość produktu z danej partii.
        Np. "worek 100 kg kukurydzy z partii KUKURYDZA-2026-001"
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - in: query
          name: batchId
          description: Filter by batch ID
          schema:
            type: string
            format: uuid
        - in: query
          name: productId
          description: Filter by product type ID
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - in: query
          name: after
          schema:
            type: string
      responses:
        '200':
          description: Inventory instance list (paginated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryInstanceListResponse'
        '401':
          description: Authentication required
    post:
      operationId: createInventoryInstance
      tags:
        - Inventory
      summary: Create an inventory instance for auction
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryInstanceRequest'
      responses:
        '201':
          description: Instance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryInstanceResponse'
        '400':
          description: Invalid request
  /api/inventory/instances/{instanceId}:
    get:
      operationId: getInventoryInstance
      tags:
        - Inventory
      summary: Get inventory instance by ID
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - in: path
          name: instanceId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Instance details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryInstanceResponse'
        '404':
          description: Instance not found
...
components:
  parameters:
    ApiVersionHeader:
      in: header
      name: X-API-Version
      required: false
      schema:
        type: string
        default: "1"
      description: API version negotiated via HTTP header.
  schemas:
...
    CreateBatchRequest:
      type: object
      required:
        - productId
        - name
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        name:
          type: string
          example: "POTATO-2024-WARSZAWA-001"
        quantity:
          type: object
          required:
            - amount
            - unit
          properties:
            amount:
              type: string
              example: "2000"
            unit:
              type: string
              example: "kg"
        dateProduced:
          type: string
          format: date-time
        bestBefore:
          type: string
          format: date-time
    BatchResponse:
      type: object
      required:
        - batchId
        - productId
        - name
        - quantity
      properties:
        batchId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: object
          properties:
            amount:
              type: string
            unit:
              type: string
        dateProduced:
          type: string
          format: date-time
        bestBefore:
          type: string
          format: date-time
    CreateInventoryInstanceRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        batchId:
          type: string
          format: uuid
        serialNumber:
          type: string
          example: "VIN-WVWZZZ3CZWE123456"
        quantity:
          type: object
          required:
            - amount
            - unit
          properties:
            amount:
              type: string
              example: "100"
            unit:
              type: string
              example: "kg"
    InventoryInstanceResponse:
      type: object
      required:
        - instanceId
        - productId
        - quantity
      properties:
        instanceId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        batchId:
          type: string
          format: uuid
        serialNumber:
          type: string
        quantity:
          type: object
          properties:
            amount:
              type: string
            unit:
              type: string
...
    BatchListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          description: List of batches
          items:
            $ref: '#/components/schemas/BatchResponse'
        nextCursor:
          type: string
          nullable: true
    InventoryInstanceListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          description: List of inventory instances
          items:
            $ref: '#/components/schemas/InventoryInstanceResponse'
        nextCursor:
          type: string
          nullable: true
...